refactor: quebrar páginas grandes em componentes/hooks + atualizar docs
Reduz o tamanho dos arquivos grandes (menos contexto por leitura) pagando as
dívidas técnicas do plano. Extrações 100% comportamento-preservantes, seguindo
o padrão orquestrador + sub-hooks + funções puras. Testids e imports mockados
por caminho preservados.

- Fase 1: mover resolveSessionConflict → sessionRecoveryService (+ teste)
- Fases 2/3: LoginPage → utils/authValidation (+ teste) e components/auth/
- Fases 4/5: ProfilePage (979→250) → components/profile/, hooks/profile/,
  utils/xpLevel (+ teste)
- Fases 6/7/8: WorkoutExecutionPage (1038→410) → components/execution/
  (TopBarButton, ExecutionTopBar, ExecutionProgressCard, FocusModeNav,
  ExecutionSkeleton, ExerciseCard unificado, ExerciseGroupCard, modais de
  cancelar/finalizar/compartilhar), hooks/workout-execution/ (navegação,
  fluxo de finalização, compartilhamento), utils/imagePreload,
  utils/shareWorkoutImage (+ teste), utils/computeSessionStats (+ teste)
- Fase 9: docs/architecture.md e README atualizados (dívida do useWorkoutSession
  paga; estratégia de custo zero das agregações client-side documentada)

npm test: 215 testes verdes (32 arquivos). npm run build: OK.
