+++ src/components/auth/GoogleIcon.jsx
@@ -1,0 +1,30 @@
+export function GoogleIcon() {
+    return (
+        <svg
+            className="w-[18px] h-[18px] shrink-0 block"
+            width="18"
+            height="18"
+            viewBox="0 0 18 18"
+            aria-hidden="true"
+            focusable="false"
+        >
+            <path
+                fill="#EA4335"
+                d="M9 3.48c1.69 0 2.84.73 3.49 1.34l2.38-2.38C13.41 1.09 11.43 0 9 0 5.48 0 2.44 2.02.96 4.95l2.77 2.15C4.45 4.91 6.5 3.48 9 3.48z"
+            />
+            <path
+                fill="#4285F4"
+                d="M17.64 9.2c0-.63-.06-1.25-.16-1.84H9v3.48h4.84a4.14 4.14 0 0 1-1.8 2.72v2.26h2.92c1.71-1.57 2.68-3.88 2.68-6.62z"
+            />
+            <path
+                fill="#FBBC05"
+                d="M3.73 10.07A5.39 5.39 0 0 1 3.43 9c0-.37.06-.72.16-1.07V5.78H.96A9 9 0 0 0 0 9c0 1.45.35 2.82.96 4.05l2.77-2.15z"
+            />
+            <path
+                fill="#34A853"
+                d="M9 18c2.43 0 4.47-.8 5.96-2.18l-2.92-2.26c-.81.54-1.85.86-3.04.86-2.5 0-4.55-1.69-5.29-3.98L.96 13.05C2.44 15.98 5.48 18 9 18z"
+            />
+            <path fill="none" d="M0 0h18v18H0z" />
+        </svg>
+    );
+}
